requirements.txt
flask
python-dotenv
openai
deepgram-sdk